+++ src/main/java/com/djarjo/common/ReflectionHelper.java
@@ -5,6 +5,7 @@ import com.google.common.flogger.FluentLogger;
 import java.lang.reflect.*;
 import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.Base64;
 import java.util.List;
 
 /**
@@ -62,9 +63,9 @@ public class ReflectionHelper {
 
 	public static Method findMethod( Method[] methods, String propertyName,
 			List<String> prefixes ) {
-		assert methods!=null : "No methods given";
+		assert methods != null : "No methods given";
 		assert !prefixes.isEmpty() : "No prefixes given";
-		if ( propertyName==null || propertyName.isEmpty() ) {
+		if ( propertyName == null || propertyName.isEmpty() ) {
 			return null;
 		}
 		String name = "" + Character.toUpperCase( propertyName.charAt( 0 ) );
@@ -189,7 +190,7 @@ public class ReflectionHelper {
 	}
 
 	public static Object getValueByIndex( Object bean, int index ) {
-		if ( bean==null ) return null;
+		if ( bean == null ) return null;
 		Class<?> beanType = bean.getClass();
 
 		if ( List.class.isAssignableFrom( beanType ) ) {
@@ -210,7 +211,7 @@ public class ReflectionHelper {
 		}
 		//--- Use getter
 		Method getter = obtainGetterFromMethod( bean.getClass(), (Method) member );
-		if ( getter==null ) {
+		if ( getter == null ) {
 			throw new RuntimeException(
 					String.format( "No getter in %s from %s", bean, member ) );
 		}
@@ -317,11 +318,39 @@ public class ReflectionHelper {
 	}
 
 	public static boolean isArrayOrList( Class<?> clazz ) {
-		return (clazz!=null) && (isList( clazz ) || clazz.isArray());
+		return (clazz != null) && (isList( clazz ) || clazz.isArray());
+	}
+
+	/**
+	 * Checks if member is a byte array
+	 *
+	 * @param member field or getter
+	 * @return {@code true} if a byte array
+	 */
+	public static boolean isByteArray( Member member ) {
+		return (member instanceof Field field) ? field.getType() == byte[].class
+				: member instanceof Method method && method.getReturnType() == byte[].class;
+	}
+
+	/**
+	 * Checks if member is an enumeration.
+	 *
+	 * @param member field or getter
+	 * @return {@code true} if an enumeration
+	 */
+	public static boolean isEnum( Member member ) {
+		if ( member instanceof Field field ) {
+			return field.getType().isEnum();
+		}
+		if ( member instanceof Method method ) {
+			Class<?> returnType = method.getReturnType();
+			return returnType.isEnum() || Enum.class.isAssignableFrom( returnType );
+		}
+		return false;
 	}
 
 	public static boolean isList( Class<?> clazz ) {
-		return clazz!=null && List.class.isAssignableFrom( clazz );
+		return clazz != null && List.class.isAssignableFrom( clazz );
 	}
 
 	/**
@@ -336,9 +365,9 @@ public class ReflectionHelper {
 	 * @return {@code true} if method is a getter
 	 */
 	public static boolean isGetter( Method method ) {
-		if ( (method==null)
-				|| (method.getParameterCount()!=0)
-				|| (method.getReturnType()==void.class)
+		if ( (method == null)
+				|| (method.getParameterCount() != 0)
+				|| (method.getReturnType() == void.class)
 				|| "getClass".equals( method.getName() ) ) {
 			return false;
 		}
@@ -364,10 +393,10 @@ public class ReflectionHelper {
 	 * @return {@code true} if method is a setter
 	 */
 	public static boolean isSetter( Method method ) {
-		if ( method==null ) return false;
+		if ( method == null ) return false;
 		String methodName = method.getName();
 		for ( String prefix : setterPrefixes ) {
-			if ( methodName.startsWith( prefix ) && method.getParameterCount()==1 ) {
+			if ( methodName.startsWith( prefix ) && method.getParameterCount() == 1 ) {
 				return true;
 			}
 		}
@@ -384,10 +413,10 @@ public class ReflectionHelper {
 	 * @return method name or {@code null} if {@code propertyName} is empty
 	 */
 	public static String makeMethodName( String prefix, String propertyName ) {
-		if ( propertyName==null || propertyName.isEmpty() ) {
+		if ( propertyName == null || propertyName.isEmpty() ) {
 			return null;
 		}
-		prefix = (prefix==null) ? "" : prefix;
+		prefix = (prefix == null) ? "" : prefix;
 		String name = prefix + Character.toUpperCase( propertyName.charAt( 0 ) );
 		if ( propertyName.length() > 1 ) {
 			name += propertyName.substring( 1 );
@@ -400,7 +429,7 @@ public class ReflectionHelper {
 			return method;
 		}
 		String name = getVarNameFromMethodName( method.getName() );
-		return (name==null) ? null : findGetter( clazz, name );
+		return (name == null) ? null : findGetter( clazz, name );
 	}
 
 	/**
@@ -414,7 +443,7 @@ public class ReflectionHelper {
 			return method;
 		}
 		String name = getVarNameFromMethodName( method.getName() );
-		return (name==null) ? null : findSetter( clazz, name );
+		return (name == null) ? null : findSetter( clazz, name );
 	}
 
 	/**
@@ -432,6 +461,7 @@ public class ReflectionHelper {
 	 */
 	public static void setValue( Object bean, Member member, int index,
 			Object value ) throws IllegalAccessException {
+
 		//--- Handle Field
 		if ( member instanceof Field field ) {
 			_setValueUsingField( bean, field, index, value );
@@ -441,7 +471,7 @@ public class ReflectionHelper {
 		//--- Handle setter method
 		if ( member instanceof Method method ) {
 			Method setter = obtainSetterFromMethod( bean.getClass(), method );
-			if ( setter==null ) {
+			if ( setter == null ) {
 				throw new RuntimeException(
 						String.format( "No setter found in bean %s from %s", bean, method ) );
 			}
@@ -452,7 +482,7 @@ public class ReflectionHelper {
 	}
 
 	private static int _getSize( Object value ) {
-		return (value==null) ? 0 : isList( value.getClass() )
+		return (value == null) ? 0 : isList( value.getClass() )
 				? ((List<?>) value).size() : (value.getClass().isArray())
 				? Array.getLength( value ) : 0;
 	}
@@ -490,7 +520,7 @@ public class ReflectionHelper {
 	 */
 	private static void _setValueAt( Object bean, Member member, int index,
 			Object value ) throws IllegalAccessException {
-		assert value!=null;
+		assert value != null;
 		Class<?>[] types = getParameterTypes( member );
 
 		//--- If value itself is a list then directly write into member
@@ -502,7 +532,7 @@ public class ReflectionHelper {
 		//--- Handle case with index
 		try {
 			Object currentValue = getValueFromMember( bean, member );
-			if ( currentValue==null ) {
+			if ( currentValue == null ) {
 				currentValue = createInstanceFromType( types[0] );
 				_writeValueToMember( bean, member, currentValue );
 			}
@@ -522,11 +552,21 @@ public class ReflectionHelper {
 
 	private static void _setValueUsingField( Object bean, Field field, int index,
 			Object value ) throws IllegalAccessException {
-		if ( value==null ) {
+		if ( value == null ) {
+			field.setAccessible( true );
 			field.set( bean, null );
 			return;
 		}
 		Class<?>[] types = getParameterTypes( field );
+
+		//--- If field is 'byte[]' and value is String then decode from Base64 to bytes
+		if ( (value instanceof String) && (types[0] == byte[].class) ) {
+			Object convertedValue = Base64.getDecoder().decode( (String) value );
+			field.setAccessible( true );
+			field.set( bean, convertedValue );
+			return;
+		}
+
 		if ( (List.class.isAssignableFrom( types[0] )) || types[0].isArray() ) {
 			_setValueAt( bean, field, index, value );
 			return;
@@ -547,11 +587,15 @@ public class ReflectionHelper {
 		}
 
 		try {
-			if ( value==null ) {
+			if ( value == null ) {
 				setter.invoke( bean, (Object) null );
 				return;
 			}
-			if ( isArrayOrList( types[0] ) ) {
+
+			//--- If field is 'byte[]' and value is String then decode from Base64 to bytes
+			if ( (value instanceof String) && (types[0] == byte[].class) ) {
+				value = Base64.getDecoder().decode( (String) value );
+			} else if ( isArrayOrList( types[0] ) ) {
 				_setValueAt( bean, setter, index, value );
 				return;
 			}